refactor pane grids
